feat: add new GetConfig command and inject ImpersonateAs variable Signed-off-by: Logan Cox <mail@logan-cox.com>
