+++ pkg/test/assert.go
@@ -5,9 +5,11 @@ import (
 	"fmt"
 	"time"
 
+	"k8s.io/client-go/discovery"
 	"sigs.k8s.io/controller-runtime/pkg/client"
 
-	"github.com/kudobuilder/kuttl/pkg/impersonation"
+	testutils "github.com/kudobuilder/kuttl/pkg/test/utils"
+	"github.com/kudobuilder/kuttl/pkg/k8s"
 )
 
 // Assert checks all provided assert files against a namespace.  Upon assert failure, it prints the failures and returns an error
@@ -25,8 +27,8 @@ func Assert(namespace string, timeout int, assertFiles ...string) error {
 	// feels like the wrong abstraction, need to do some refactoring
 	s := &Step{
 		Timeout:         0,
-		Client:          impersonation.Client,
-		DiscoveryClient: impersonation.DiscoveryClient,
+		Client:          Client,
+		DiscoveryClient: DiscoveryClient,
 	}
 
 	var testErrors []error
@@ -70,8 +72,8 @@ func Errors(namespace string, timeout int, errorFiles ...string) error {
 	// feels like the wrong abstraction, need to do some refactoring
 	s := &Step{
 		Timeout:         0,
-		Client:          impersonation.Client,
-		DiscoveryClient: impersonation.DiscoveryClient,
+		Client:          Client,
+		DiscoveryClient: DiscoveryClient,
 	}
 
 	var testErrors []error
@@ -102,3 +104,37 @@ func Errors(namespace string, timeout int, errorFiles ...string) error {
 	return errors.New("error asserts not valid")
 }
 
+
+func Client(_ bool) (client.Client, error) {
+	cfg, err := k8s.GetConfig()
+
+	if err != nil {
+		return nil, err
+	}
+
+	client, err := testutils.NewRetryClient(cfg, client.Options{
+		Scheme: testutils.Scheme(),
+	})
+
+	if err != nil {
+		return nil, fmt.Errorf("fatal error getting client: %v", err)
+	}
+
+	return client, nil
+}
+
+func DiscoveryClient() (discovery.DiscoveryInterface, error) {
+	cfg, err := k8s.GetConfig()
+
+	if err != nil {
+		return nil, err
+	}
+
+	dclient, err := discovery.NewDiscoveryClientForConfig(cfg)
+
+	if err != nil {
+		return nil, fmt.Errorf("fatal error getting discovery client: %v", err)
+	}
+
+	return dclient, nil
+}
